voicewave-backend/requirements.txt
aiohttp==3.11.18
aiohttp-cors==0.8.1
annotated-types==0.7.0
fastapi==0.115.12
h11==0.16.0
huggingface-hub==0.30.2
librosa==0.11.0
numpy==1.26.4
openai-whisper==20240930
pydantic==2.11.3
python-multipart==0.0.20
ray[serve]==2.44.1
requests==2.32.3
scipy==1.12.0
soundfile==0.13.1
starlette==0.46.2
torch==2.7.0
transformers==4.51.3
tqdm==4.66.2
typing_extensions==4.13.2
uvicorn==0.34.2
websockets==15.0.1